+++ cypress/support/index.js
@@ -19,3 +19,5 @@ import "./commands";
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
 require("cypress-dark");
+
+import "@cypress/code-coverage/support";