+++ src/contexts/AccountContext/types.ts
@@ -1,5 +1,5 @@
 export interface ContextValues {
   accountRs?: string;
   accountSeed?: string;
-  fetchFn?: Function;
+  fetchFn?: () => Promise<void>;
 }